rename unified ViewController to AddGroupMembersViewController
